verify all jerasure variants
The code path used for encoding / decoding in the jerasure plugin
depends on the available CPU features. Detect which variants of the
jerasure plugin could be used for encoding / decoding on the host
running the test and try all of them. This verifies that all variants
indeed encode / decode exactly in the same way. It is a safeguard
against a regression / corruption in case the a bug is introduced in the
SSE4 variant that would make it encode differently than the SSE3
variant.

http://tracker.ceph.com/issues/9720 Fixes: #9720

Signed-off-by: Loic Dachary <loic@dachary.org>