verify all payloads have been verified
Collect the directories containing payloads to be verified. At the end
of the test, if all tests have run, verify that all directories
containing payload have been used. If not it probably means the tests
were changed and no longer use these directories. They should be
removed. It's also a guard against unintended regression that skips
payloads that should be verified.

Signed-off-by: Loic Dachary <loic@dachary.org>