Test with new gf-complete
With the new support for SIMD runtime detection in gf-complete,
the non-regression tests have been modified to test with some of
the SIMD detection disabled.

All shec and jerasure tests now run with different SIMD support
turned on and off. This will ensure the decoding the corpus
with code paths that are mathematically equivalent but optimized
for different processors.

Signed-off-by: Bassam Tabbara <bassam.tabbara@quantum.com>